First attempt at passing hands into methods and not values
I think that it doesn't make sense to call these methods sometimes on a hand and sometimes on a list of values. It makes more sense to pass around the same kind of 'object'.

Right now, changing the way the two of a kind method works will break the implementation of the two pair method. I've left it as is for now.